+++ frontend/wailsjs/go/app/App.js
@@ -846,6 +846,14 @@ export function StartOAuthFlow(arg1) {
   return window['go']['app']['App']['StartOAuthFlow'](arg1);
 }
 
+export function SubscribeAllFolders(arg1) {
+  return window['go']['app']['App']['SubscribeAllFolders'](arg1);
+}
+
+export function SubscribeFolder(arg1, arg2) {
+  return window['go']['app']['App']['SubscribeFolder'](arg1, arg2);
+}
+
 export function SyncAccountComplete(arg1) {
   return window['go']['app']['App']['SyncAccountComplete'](arg1);
 }
@@ -900,6 +908,10 @@ export function Undo() {
 
 export function Unstar(arg1) {
   return window['go']['app']['App']['Unstar'](arg1);
+}
+
+export function UnsubscribeFolder(arg1, arg2) {
+  return window['go']['app']['App']['UnsubscribeFolder'](arg1, arg2);
 }
 
 export function UpdateAccount(arg1, arg2) {